Gitlab-CI continuous integration
This commit adds the necessary files to run the CI using Gitlab-CI. For
now, it checks that Weboob builds, then runs the linting script
(checking that every module as an icon and some tests + PyFlakes) and
the unittests. Most of modules unittests cannot run because there is no
backend configured.

Some changes were needed in the pre-existing scripts:
* Edit `weboob_lint` to exit with non-zero code if it finds modules
without icons or tests, so that the build could fail in such a case.
* Edit `run_tests.sh` to set correct exit code on failure and rework
generation of XUNIT output. Also added some doc about useful environment
variables. Added a way to generate an xunit output file when running modules
unittests, passing a `XUNIT_OUT` env variable to `run_tests.sh` script.
* Modification of `setup.cfg` and `run_tests` scripts to handle code
coverage generation. The matching regex in Gitlab for the total code
coverage is `TOTAL: (\d+\%\s*)$)`.

I also added a script to generate a JSON module status matrix from
modules unittests, ready to be sent to a
[Weboob-CI](https://github.com/Phyks/weboob-ci) instance.

NOTE: Required Python modules are taken from the `setup.py` script.
`.ci/requirements.txt` contains the requirements to run the unittests
and the CI, whereas `.ci/requirements_modules.txt` contains the specific
Python modules required at runtime by Weboob modules. The latter could
eventually be replaced by a proper call to `debpydep` script.

+++ tools/run_tests.sh
@@ -1,4 +1,9 @@
-#!/bin/sh
+#!/bin/bash
+
+# Mai available environment variables
+#   * RSYNC_TARGET: target on which to rsync the xunit output.
+#   * XUNIT_OUT: file in which xunit output should be saved.
+#   * WEBOOB_BACKENDS: path to the Weboob backends file to use.
 
 # stop on failure
 set -e
@@ -30,6 +35,10 @@ else
     RSYNC_TARGET=""
 fi
 
+if [ ! -n "${XUNIT_OUT}" ]; then
+    XUNIT_OUT=""
+fi
+
 # find executables
 if [ -z "${PYTHON}" ]; then
     which python >/dev/null 2>&1 && PYTHON=$(which python)
@@ -56,11 +65,18 @@ fi
 # do not allow undefined variables anymore
 set -u
 WEBOOB_TMPDIR=$(mktemp -d "${TMPDIR}/weboob_test.XXXXX")
-cp "${WEBOOB_BACKENDS}" "${WEBOOB_TMPDIR}/backends"
+if [ -f "${WEBOOB_BACKENDS}" ]; then
+    cp "${WEBOOB_BACKENDS}" "${WEBOOB_TMPDIR}/backends"
+else
+    touch "${WEBOOB_TMPDIR}/backends"
+    chmod go-r "${WEBOOB_TMPDIR}/backends"
+fi
 
 # xunit nose setup
 if [ -n "${RSYNC_TARGET}" ]; then
     XUNIT_ARGS="--with-xunit --xunit-file=${WEBOOB_TMPDIR}/xunit.xml"
+elif [ -n "${XUNIT_OUT}" ]; then
+    XUNIT_ARGS="--with-xunit --xunit-file=${XUNIT_OUT}"
 else
     XUNIT_ARGS=""
 fi
@@ -77,17 +93,40 @@ ${PYTHON} "${WEBOOB_DIR}/scripts/weboob-config" update
 
 # allow failing commands past this point
 set +e
+set -o pipefail
 if [ -n "${BACKEND}" ]; then
     ${PYTHON} ${NOSE} -c /dev/null -sv "${WEBOOB_MODULES}/${BACKEND}/test.py" ${XUNIT_ARGS}
     STATUS=$?
     STATUS_CORE=0
 else
     echo "=== Weboob ==="
-    ${PYTHON} ${NOSE} -c ${WEBOOB_DIR}/setup.cfg -sv
+    CORE_TESTS=$(mktemp)
+    ${PYTHON} ${NOSE} --cover-package weboob -c ${WEBOOB_DIR}/setup.cfg -sv 2>&1 | tee "${CORE_TESTS}"
     STATUS_CORE=$?
     echo "=== Modules ==="
-    find "${WEBOOB_MODULES}" -name "test.py" | sort | xargs ${PYTHON} ${NOSE} -c /dev/null -sv ${XUNIT_ARGS}
+    MODULES_TESTS=$(mktemp)
+    MODULES_TO_TEST=$(find "${WEBOOB_MODULES}" -name "test.py" | sort | xargs echo)
+    ${PYTHON} ${NOSE} --with-coverage --cover-package modules -c /dev/null -sv ${XUNIT_ARGS} ${MODULES_TO_TEST} 2>&1 | tee ${MODULES_TESTS}
     STATUS=$?
+
+    # Compute total coverage
+    echo "=== Total coverage ==="
+    CORE_STMTS=$(grep "TOTAL" ${CORE_TESTS} | awk '{ print $2; }')
+    CORE_MISS=$(grep "TOTAL" ${CORE_TESTS} | awk '{ print $3; }')
+    CORE_COVERAGE=$(grep "TOTAL" ${CORE_TESTS} | awk '{ print $4; }')
+    MODULES_STMTS=$(grep "TOTAL" ${MODULES_TESTS} | awk '{ print $2; }')
+    MODULES_MISS=$(grep "TOTAL" ${MODULES_TESTS} | awk '{ print $3; }')
+    MODULES_COVERAGE=$(grep "TOTAL" ${MODULES_TESTS} | awk '{ print $4; }')
+    echo "CORE COVERAGE: ${CORE_COVERAGE}"
+    echo "MODULES COVERAGE: ${MODULES_COVERAGE}"
+    TOTAL_STMTS=$((${CORE_STMTS} + ${MODULES_STMTS}))
+    TOTAL_MISS=$((${CORE_MISS} + ${MODULES_MISS}))
+    TOTAL_COVERAGE=$((100 * (${TOTAL_STMTS} - ${TOTAL_MISS}) / ${TOTAL_STMTS}))
+    echo "TOTAL: ${TOTAL_COVERAGE}%"
+
+    # removal of temp files
+    rm ${CORE_TESTS}
+    rm ${MODULES_TESTS}
 fi
 
 # xunit transfer
